requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
websockets==14.2
python-dotenv==1.0.1
google-generativeai==0.8.4
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.2.2
sqlalchemy==2.0.38
